requirements.txt
boto3==1.15.16
botocore==1.18.16
Django==3.1.2
django-anymail==8.1
django-cors-headers==3.5.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
django-simple-history==2.11.0
django-storages==1.10.1
django-templated-email==2.3.0
djangorestframework==3.12.1
djoser==2.0.5
mysqlclient==2.0.1
Pillow==7.2.0
requests==2.24.0
stripe==2.54.0
